requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.0
asttokens==2.2.1
attrs==25.3.0
autopep8==2.3.2
backcall==0.2.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.2.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
dice==4.0.0
distro==1.9.0
exceptiongroup==1.2.2
executing==1.2.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.0.3
jupyter_core==5.2.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.6
narwhals==2.12.0
nest-asyncio==1.5.6
numpy==2.2.5
openai==2.8.1
outcome==1.3.0.post0
packaging==23.0
pandas==2.2.3
parso==0.8.3
pickleshare==0.7.5
pillow==12.0.0
platformdirs==3.1.1
pluggy==1.6.0
prompt-toolkit==3.0.38
protobuf==6.33.1
psutil==5.9.4
pure-eval==0.2.2
pyarrow==21.0.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyfiglet==0.8.post1
Pygments==2.14.0
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.8.2
python-dotenv==1.2.1
pytz==2025.2
pywin32==305
PyYAML==6.0.2
pyzmq==25.0.0
random_word==1.0.13
referencing==0.37.0
requests==2.32.3
rpds-py==0.29.0
selenium==4.31.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.2
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
tomli==2.2.1
tornado==6.2
tqdm==4.67.1
traitlets==5.9.0
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.6
websocket-client==1.8.0
wsproto==1.2.0
